+++ transitions/extensions/__init__.py
@@ -7,10 +7,16 @@ and threadsafe execution of machine methods. Additionally, combinations of all t
 and made easier to access with a convenience factory.
 """
 
-from .diagrams import GraphMachine
+from .diagrams import GraphMachine, HierarchicalGraphMachine
 from .nesting import HierarchicalMachine
 from .locking import LockedMachine
 
-from .factory import MachineFactory, HierarchicalGraphMachine, LockedHierarchicalGraphMachine
+from .factory import MachineFactory, LockedHierarchicalGraphMachine
 from .factory import LockedHierarchicalMachine, LockedGraphMachine
-from .factory import AsyncMachine, AsyncGraphMachine, HierarchicalAsyncMachine, HierarchicalAsyncGraphMachine
+
+try:
+    # only available for Python 3
+    from .asyncio import AsyncMachine, HierarchicalAsyncMachine
+    from .factory import AsyncGraphMachine, HierarchicalAsyncGraphMachine
+except (ImportError, SyntaxError):
+    pass