+++ transitions/extensions/markup.py
@@ -14,7 +14,19 @@ import numbers
 
 from six import string_types, iteritems
 
-from ..core import Machine, Enum
+try:
+    # Enums are supported for Python 3.4+ and Python 2.7 with enum34 package installed
+    from enum import Enum, EnumMeta
+except ImportError:
+    # If enum is not available, create dummy classes for type checks
+    # typing must be prevent redefinition issues with mypy
+    class Enum:  # type:ignore
+        """ This is just an Enum stub for Python 2 and Python 3.3 and before without Enum support. """
+
+    class EnumMeta:  # type:ignore
+        """ This is just an EnumMeta stub for Python 2 and Python 3.3 and before without Enum support. """
+
+from ..core import Machine
 from .nesting import HierarchicalMachine